backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
google-generativeai
python-multipart
